// backend/src/modules/user/enums/platform-permission.enum.ts
export enum PlatformPermissionEnum {
    LISTING_CREATE = "PLATFORM:LISTING:CREATE",
    LISTING_EDIT = "PLATFORM:LISTING:EDIT",
    LISTING_EDIT_BY_MANAGER = "PLATFORM:LISTING:EDIT:BY_MANAGER",
    LISTING_EDIT_BY_ADMIN = "PLATFORM:LISTING:EDIT:BY_ADMIN",
    LISTING_DELETE = "PLATFORM:LISTING:DELETE",
    LISTING_ACTIVATE = "PLATFORM:LISTING:ACTIVATE",
    LISTING_DEACTIVATE = "PLATFORM:LISTING:DEACTIVATE",
    LISTING_GET_STATS = "PLATFORM:LISTING:GET_STATS",
    LISTING_CLOSE = "PLATFORM:LISTING:CLOSE",
    LISTING_GET_ALL_FOR_STAFF = "PLATFORM:LISTING:GET_ALL:FOR_STAFF",
    LISTING_GET_BY_ID_FOR_STAFF = "PLATFORM:LISTING:GET:BY_ID:FOR_STAFF",
    LISTING_GET_MY_ALL = "PLATFORM:LISTING:GET:MY_ALL",
    LISTING_GET_MY = "PLATFORM:LISTING:GET:MY",
    LISTING_UPLOAD_POSTER = "PLATFORM:LISTING:UPLOAD_POSTER",
    LISTING_DELETE_POSTER = "PLATFORM:LISTING:DELETE_POSTER",

    USER_GET_ALL_FOR_STAFF = "PLATFORM:USER:GET_ALL:FOR_STAFF",
    USER_GET_BY_ID_FOR_STAFF = "PLATFORM:USER:GET:BY_ID:FOR_STAFF",
    USER_EDIT_BY_ADMIN = "PLATFORM:USER:EDIT:BY_ADMIN",
    USER_ACTIVATE = "PLATFORM:USER:ACTIVATE",
    USER_DEACTIVATE = "PLATFORM:USER:DEACTIVATE",
    USER_DELETE = "PLATFORM:USER:DELETE",
    USER_SET_ROLE = "PLATFORM:USER:SET_ROLE",

    ME_GET = "PLATFORM:ME:GET",
    ME_EDIT = "PLATFORM:ME:EDIT",
    ME_TOP_UP = "PLATFORM:ME:TOP_UP",
    ME_BECOME_SELLER = "PLATFORM:ME:BECOME_SELLER",
    ME_UPGRADE_PLAN = "PLATFORM:ME:UPGRADE_PLAN",
    ME_DEACTIVATE = "PLATFORM:ME:DEACTIVATE",
    ME_LISTING_DEACTIVATE = "PLATFORM:ME:LISTING:DEACTIVATE",
    ME_UPLOAD_AVATAR = "PLATFORM:ME:UPLOAD_AVATAR",
    ME_DELETE_AVATAR = "PLATFORM:ME:DELETE_AVATAR",

    SUBSCRIPTION_SET_PLAN = "PLATFORM:SUBSCRIPTION:SET:PLAN",
    SUBSCRIPTION_ACTIVATE = "PLATFORM:SUBSCRIPTION:ACTIVATE",
    SUBSCRIPTION_DEACTIVATE = "PLATFORM:SUBSCRIPTION:DEACTIVATE",

    CAR_MISSING_REPORT = "PLATFORM:CAR:MISSING_REPORT",
}